+++ django_spire/notification/sms/admin.py
@@ -1,10 +1,17 @@
 from django.contrib import admin
 
-from django_spire.notification.sms.models import SmsNotification
+from django_spire.notification.sms.models import SmsNotification, SmsTemporaryMedia
 
 
 @admin.register(SmsNotification)
-class EmailNotificationAdmin(admin.ModelAdmin):
+class SmsNotificationAdmin(admin.ModelAdmin):
     list_display = (
         'id', 'notification', 'to_phone_number'
+    )
+
+
+@admin.register(SmsTemporaryMedia)
+class SmsTemporaryMediaAdmin(admin.ModelAdmin):
+    list_display = (
+        'id', 'name', 'content_type', 'external_url'
     )