+++ django_spire/notification/sms/urls/__init__.py
@@ -1,0 +1,8 @@
+from django.urls import include, path
+
+app_name = 'sms'
+
+
+urlpatterns = [
+    path('media/', include('django_spire.notification.sms.urls.media_urls', namespace='media')),
+]