+++ django_spire/contrib/service/__init__.py
@@ -1,0 +1,7 @@
+from django_spire.contrib.service.service import BaseService
+from django_spire.contrib.service.django_model_service import BaseDjangoModelService
+
+__all__ = [
+    'BaseService',
+    'BaseDjangoModelService'
+]