+++ docs/changelog/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.9.1
+
+### Features
+
+- Added content object querysets to all `Notification` queryset managers
+
 ## v0.9.0
 
 ### Features